moves docking and preparation functions into seperate modules
